instoreitem

+++ src/main/java/dhu/sun/vege/entity/InstoreItem.java
@@ -26,13 +26,13 @@ public class InstoreItem extends BaseEntity{
     private String name;
 
     @Column(name = "number")
-    private int number;
+    private Double number;
 
     @Column(name = "price")
-    private BigDecimal price;
+    private Double price;
 
     @Column(name = "total")
-    private BigDecimal total;
+    private Double total;
 
     public Long getId() {
         return id;
@@ -66,27 +66,27 @@ public class InstoreItem extends BaseEntity{
         this.name = name;
     }
 
-    public int getNumber() {
+    public Double getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(Double number) {
         this.number = number;
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
 
-    public BigDecimal getTotal() {
+    public Double getTotal() {
         return total;
     }
 
-    public void setTotal(BigDecimal total) {
+    public void setTotal(Double total) {
         this.total = total;
     }
 }
